pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dragoneye-auditor"
version = "1.0.0"
description = "OpenEnv environment for auditing Hinglish e-commerce reviews."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.135.2",
    "uvicorn>=0.42.0",
    "pydantic>=2.12.5",
    "openai>=2.30.0",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
    "openenv-core>=0.2.3"
]

[project.urls]
"Homepage" = "https://huggingface.co/spaces/"

[project.scripts]
server = "server.app:main"